requirements.txt
asgiref==3.5.2
attrs==21.4.0
autobahn==22.4.2
Automat==20.2.0
certifi==2022.5.18.1
cffi==1.15.0
channels==3.0.4
charset-normalizer==2.0.12
constantly==15.1.0
cryptography==37.0.2
daphne==3.0.2
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2
django-allauth==0.50.0
django-cities==0.6
django-countries==7.3.2
django-crispy-forms==1.14.0
django-model-utils==4.2.0
django-phone-field==1.8.1
django-rest-framework==0.1.0
djangorestframework==3.13.1
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
oauthlib==3.2.0
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.4.0
pyOpenSSL==22.0.0
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
swapper==1.3.0
tqdm==4.64.0
twilio==7.9.2
Twisted==22.4.0
txaio==22.2.1
typing-extensions==4.2.0
urllib3==1.26.9
whitenoise==6.1.0
zope.interface==5.4.0
